ai-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML - Multiple provider support
google-generativeai==0.3.2  # Primary: Gemini API
openai==1.3.7               # Optional: OpenAI GPT
anthropic==0.7.7            # Optional: Claude

# Audio processing
openai-whisper==20231117
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# Image processing
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3

# PDF and document processing
pypdf==3.17.1
python-docx==1.1.0
pdfplumber==0.10.3

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
newspaper3k==0.2.8

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2

# Video processing (optional)
moviepy==1.0.3
youtube-transcript-api==0.6.1

# Additional dependencies for Phase 2
PyPDF2==3.0.1  # Backup PDF reader
